changed endpoint to /hosts/n/reservations and moved it
